Fix compilation on macOS Nix clang
Fixes 'error: scalar initializer cannot be empty'

The '= {}' empty initializer is a GNU extension, not supported on older
versions of clang. For arrays, we can use '= {0}' to ensure they are
zero-initialized. When using a designated initializer, unspecified
members are zero-initialized, so we can simply omit them.

To prevent regressions, recent clang can warn on these with
-Wgnu-empty-initializer, so this is enabled if supported.
However, GNU would need -Wpedantic to do the same.

Signed-off-by: Ryan Burns <rtburns@protonmail.com>